created string_view.h
